프로젝트 연동/소스 등록
프젝ID값으로 소스리스트 및 설정 판별가능

+++ src/main/java/egovframework/example/vo/SourceVO.java
@@ -14,5 +14,5 @@ public class SourceVO {
 	private LocalDate createdAt;		// 생성일자
 	private boolean isDeleted;			// 삭제
 	
-	private int projectId;
+	private Long projectId;
 }
